Move pkg/tfbridge/validate_input_types.go to pkg/tfbridge/typechecker/typechecker.go
This is an isolated set of functionality, so it should be in its own package.
